Only clean up temp dir in setup if it exists

+++ geosoft/gxpy/tests/__init__.py
@@ -85,7 +85,8 @@ class GXPYTest(unittest.TestCase):
         gxu._temp_folder_override = os.path.join(cls._result_base_dir, '__tmp__')
         # TODO: For some reason some tests can leave files in temp. Clean up now to keep tests stable,
         #       but investigate the root of the issue
-        shutil.rmtree(gxu._temp_folder_override)
+        if os.path.exists(gxu._temp_folder_override):
+            shutil.rmtree(gxu._temp_folder_override)
         os.makedirs(gxu._temp_folder_override, exist_ok=True)
 
         gxu._uuid_callable = cls._cls_uuid
